add slider control to main library
