+++ src/render.js
@@ -5,7 +5,7 @@ import ticked from './ticked';
 import dragstarted from './dragstarted';
 import dragged from './dragged';
 import dragended from './dragended';
-import drawHelpText from './drawHelpText';
+import drawText from './drawText';
 
 export default function render(props) {
   //
@@ -362,9 +362,24 @@ export default function render(props) {
   );
 
   // draw the help text
-  drawHelpText({
+  drawText({
     selector: 'svg',
+    text: `mouse over a node to see it's relationships. click the background to reset.`,
     height
+  });
+
+  d3
+    .select('div#graph')
+    .append('div')
+    .attr('id', 'slider-container');
+
+  // draw the slider control
+  drawSliderControl({
+    selector: 'div#slider-container',
+    padding: '10px',
+    defaultLinkOpacity: 0.4,
+    defaultMarkOpacity: 0.4,
+    defaultLabelOpacity: 1
   });
 
   //
@@ -540,5 +555,68 @@ export default function render(props) {
         .selectAll('.link')
         .style('stroke-opacity', defaultLinkOpacity);
     };
+  }
+
+  /* global d3 */
+  function drawSliderControl(props) {
+    var selector = props.selector;
+    var padding = props.padding;
+    var defaultMarkOpacity = props.defaultMarkOpacity;
+    var defaultLinkOpacity = props.defaultLinkOpacity;
+    var defaultLabelOpacity = props.defaultLabelOpacity;
+
+    d3
+      .select(selector)
+      .append('input')
+      .attr('type', 'range')
+      .attr('min', 0)
+      .attr('max', 1)
+      .attr('value', 0.356)
+      .attr('step', 0.001)
+      .style('top', '604px')
+      .style('left', '90px')
+      .style('height', '36px')
+      .style('width', '450px')
+      .style('position', 'fixed')
+      .attr('id', 'slider');
+
+    d3.select('#slider').on('input', function() {
+      update(+this.value);
+    });
+
+    function update(sliderValue) {
+      console.log('sliderValue', sliderValue);
+      // adjust the text on the range slider
+      d3.select('#nRadius-value').text(sliderValue);
+      d3.select('#nRadius').property('value', sliderValue);
+
+      d3.selectAll('.link').style('stroke-opacity', function(d) {
+        // console.log('d from slider update', d);
+        if (d.weight < sliderValue) {
+          return 0;
+        }
+        return defaultLinkOpacity;
+      });
+
+      d3.selectAll('.mark').style('fill-opacity', function(d) {
+        // first style the label associated with the mark
+        // console.log('d from mark selection', d);
+        d3
+          .select('#node' + d.id)
+          .selectAll('.label')
+          .style('fill-opacity', function() {
+            if (d.maxLinkWeight < sliderValue) {
+              return 0.1;
+            }
+            return defaultLabelOpacity;
+          });
+
+        // then style the mark itself
+        if (d.maxLinkWeight < sliderValue) {
+          return 0.1;
+        }
+        return defaultMarkOpacity;
+      });
+    }
   }
 }
